refactor: extract streams send/recv to stream.rs, send_seam to its own file (Phase 5)
Move Group F (H3Stream/H3SendStream/H3RecvStream, Send/Recv/Open exec seams,
submit_owned_send, receive budget, stream ctx + callback, all stream h3 trait
impls) out of lib.rs into stream.rs, co-locating the receive_events tests.
H3*Stream types re-exported with pub use (public paths preserved). The send_seam
test module moves to a crate-root send_seam.rs so its CI --exact test path is
preserved. Pure move; no behavior/API change. 176 tests / 2 ignored; --exact
gate 4 passed. lib.rs now 943 lines (from 7679).
